refactor: consolidate frontend into single interface app
