--- backend/static/js/utils.js
@@ -1,32 +1,0 @@
-export const utils = {
-    debounce(func, wait) {
-        let timeout;
-        return function executedFunction(...args) {
-            const later = () => {
-                clearTimeout(timeout);
-                func(...args);
-            };
-            clearTimeout(timeout);
-            timeout = setTimeout(later, wait);
-        };
-    },
-
-    generateId() {
-        return 'id_' + Math.random().toString(36).substr(2, 9) + '_' + Date.now();
-    },
-
-    sanitizeHtml(str) {
-        const div = document.createElement('div');
-        div.textContent = str;
-        return div.innerHTML;
-    },
-
-    escapeRegex(str) {
-        return str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    },
-
-    formatPercent(value) {
-        return (value * 100).toFixed(2) + '%';
-    }
-};
-
