+++ backend/templates/interface.html
@@ -62,7 +62,6 @@
         </div>
     </div>
 
-    <script src="/static/js/config.js"></script>
     <script src="https://unpkg.com/vue@3/dist/vue.global.prod.js"></script>
     <script src="https://unpkg.com/pinia@2/dist/pinia.iife.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.11.174/pdf.min.js"></script>
